requirements.txt
streamlit
PyMuPDF
Pillow
opencv-python-headless
numpy
pandas
openpyxl
